chore(presentation): restructure directory
