+++ presentation/src/main/java/com/example/alpacamusclemaintenance/presentation/excercise/PushUpFragment.kt
@@ -1,4 +1,4 @@
-package com.example.alpacamusclemaintenance.presentation.ui
+package com.example.alpacamusclemaintenance.presentation.excercise
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,6 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import com.example.alpacamusclemaintenance.presentation.databinding.FragmentPushUpBinding
-import com.example.alpacamusclemaintenance.presentation.viewmodel.PushUpViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
