+++ presentation/src/main/java/com/example/alpacamusclemaintenance/presentation/record/RecordFragment.kt
@@ -1,4 +1,4 @@
-package com.example.alpacamusclemaintenance.presentation.ui
+package com.example.alpacamusclemaintenance.presentation.record
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,7 +8,6 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import com.example.alpacamusclemaintenance.domain.pushup.PushUp
 import com.example.alpacamusclemaintenance.presentation.databinding.FragmentRecordBinding
-import com.example.alpacamusclemaintenance.presentation.viewmodel.RecordViewModel
 import com.github.mikephil.charting.animation.Easing
 import com.github.mikephil.charting.charts.BarChart
 import com.github.mikephil.charting.components.XAxis
