+++ presentation/src/main/res/layout/fragment_exercise.xml
@@ -7,7 +7,7 @@
   <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.ExerciseFragment">
+    tools:context=".excercise.ExerciseFragment">
 
     <com.google.android.material.tabs.TabLayout
       android:id="@+id/tab"
