+++ presentation/src/main/res/layout/fragment_home.xml
@@ -14,7 +14,7 @@
   <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.HomeFragment">
+    tools:context=".home.HomeFragment">
 
     <TextView
       android:id="@+id/word_of_wisdom"
