+++ presentation/src/main/res/layout/fragment_push_up.xml
@@ -22,7 +22,7 @@
   <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.PushUpFragment">
+    tools:context=".excercise.PushUpFragment">
 
     <TextView
       android:id="@+id/textView"
