+++ presentation/src/main/res/layout/fragment_record.xml
@@ -14,7 +14,7 @@
   <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.RecordFragment">
+    tools:context=".record.RecordFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
       android:id="@+id/linearLayout"
