+++ presentation/src/main/res/layout/fragment_squat.xml
@@ -5,7 +5,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".ui.SquatFragment">
+  tools:context=".excercise.SquatFragment">
 
   <!-- TODO: Update blank fragment layout -->
   <TextView
