+++ presentation/src/main/res/layout/fragment_web_view.xml
@@ -5,7 +5,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".ui.WebViewFragment">
+  tools:context=".WebViewFragment">
 
   <WebView
     android:id="@+id/webView"
