+++ presentation/src/main/res/navigation/nav_graph.xml
@@ -7,25 +7,25 @@
 
   <fragment
     android:id="@+id/homeFragment"
-    android:name="com.example.alpacamusclemaintenance.presentation.ui.HomeFragment"
+    android:name="com.example.alpacamusclemaintenance.presentation.home.HomeFragment"
     android:label="@string/title_home"
     tools:layout="@layout/fragment_home" />
 
   <fragment
     android:id="@+id/exerciseFragment"
-    android:name="com.example.alpacamusclemaintenance.presentation.ui.ExerciseFragment"
+    android:name="com.example.alpacamusclemaintenance.presentation.excercise.ExerciseFragment"
     android:label="@string/title_exercise"
     tools:layout="@layout/fragment_exercise" />
 
   <fragment
     android:id="@+id/recordFragment"
-    android:name="com.example.alpacamusclemaintenance.presentation.ui.RecordFragment"
+    android:name="com.example.alpacamusclemaintenance.presentation.record.RecordFragment"
     android:label="@string/title_record"
     tools:layout="@layout/fragment_record" />
 
   <fragment
     android:id="@+id/feedFragment"
-    android:name="com.example.alpacamusclemaintenance.presentation.ui.FeedFragment"
+    android:name="com.example.alpacamusclemaintenance.presentation.feed.FeedFragment"
     android:label="@string/title_feed"
     tools:layout="@layout/fragment_feed">
 
@@ -36,7 +36,7 @@
 
   <fragment
     android:id="@+id/webViewFragment"
-    android:name="com.example.alpacamusclemaintenance.presentation.ui.WebViewFragment"
+    android:name="com.example.alpacamusclemaintenance.presentation.WebViewFragment"
     android:label="WebViewFragment"
     tools:layout="@layout/fragment_web_view">
 
@@ -45,7 +45,7 @@
 
   <fragment
     android:id="@+id/bugReportFragment"
-    android:name="com.example.alpacamusclemaintenance.presentation.ui.WebViewFragment"
+    android:name="com.example.alpacamusclemaintenance.presentation.WebViewFragment"
     android:label="WebViewFragment"
     tools:layout="@layout/fragment_web_view">
